A4#1-Removed unused class ...
Warning : to see below

http://stackoverflow.com/questions/16918826/objective-c-implicit-convers
ion-loses-integer-precision-nsuinteger-aka-unsig

+++ Matchismo/SettingsGameViewController.m
@@ -2,8 +2,8 @@
 //  SettingsGameViewController.m
 //  Matchismo
 //
-//  Created by dac duy nguyen on 4/21/14.
-//  Copyright (c) 2014 dac duy nguyen. All rights reserved.
+//  Created 4/21/14.
+//  
 //
 
 #import "SettingsGameViewController.h"
@@ -33,10 +33,10 @@
 
 -(void)setLabel:(UILabel *)label forSlider:(UISlider *)slider
 {
-    int sliderValue;
+    NSUInteger sliderValue;
     sliderValue = lround(slider.value);
     [slider setValue:sliderValue animated:NO];
-    label.text = [NSString stringWithFormat:@"%d",sliderValue];
+    label.text = [NSString stringWithFormat:@"%lu", (unsigned long) sliderValue];
 }
 
 - (IBAction)matchBonusSliderChanged:(UISlider *)sender {
